feat(tuning): add all

+++ Recommenders/GraphBased/RP3betaRecommender.py
@@ -111,7 +111,7 @@ class RP3betaRecommender(BaseItemSimilarityMatrixRecommender):
                     row_data = row_data[non_zero_mask]
 
                 similarity_builder.add_data_lists(
-                    row_list_to_add=np.ones(len(row_data), dtype=np.int) * (current_block_start_row + row_in_block),
+                    row_list_to_add=np.ones(len(row_data), dtype=np.int32) * (current_block_start_row + row_in_block),
                     col_list_to_add=relevant_items_partition,
                     data_list_to_add=row_data)
